- fixed typos and 64-bit compatibility - conform to idle callback functionality in devel_0_38
git-svn-id: https://svn.grrrr.org/ext/trunk@986 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ source/fldsp.cpp
@@ -89,7 +89,7 @@ void flext_dsp::dspmeth(fts_word_t *w)
 #else
 t_int *flext_dsp::dspmeth(t_int *w) 
 { 
-    flext_dsp *obj = (flext_dsp *)w[1];
+    flext_dsp *obj = (flext_dsp *)(size_t)w[1];
 /*
 #ifdef FLEXT_DEBUG
     if(!obj->thisHdr()) {
